+++ app/endpoints/todo/models.py
@@ -1,0 +1,53 @@
+from typing import List, Optional, Union, Set
+from pydantic import BaseModel, Schema, Json, UUID1
+import uuid
+from datetime import datetime, timedelta
+import json
+
+
+class ToDoBase(BaseModel):
+    # todoId: Optional[str]
+    title: str
+    description: Optional[str] = None
+    isComplete: Optional[bool] = False
+    dateDue: Optional[datetime] = None
+    dateCreate: Optional[str] = None
+    dateUpdate: datetime
+    dateComplete: Optional[datetime] = None
+    userId: str
+
+
+# Properties to receive via API on creation
+# Additional properties to return via API
+class ToDoInDB(ToDoBase):
+    todoId: str
+
+
+# Create a ToDo
+class TodoCreate(BaseModel):
+    title: str
+    description: Optional[str] = None
+    dateDue: Optional[datetime] = None
+    userId: Optional[str]
+
+
+# # Create a ToDo
+# class TodoGetId(ToDoBase):
+#     todoId: str
+
+# # Update a ToDo
+# class TodoUpdate(ToDoBase):
+#     todoId: str
+#     title: str
+#     description: str
+#     isComplete: Optional[bool] = False
+#     dateDue: Optional[str]
+
+# # complete a ToDo
+# class TodoUpdateComplete(ToDoBase):
+#     todoId: str
+#     isComplete: bool = True
+
+# # Delete a ToDo
+# class TodoDelete(ToDoBase):
+#     todoId: str